+++ repository/src/main/java/by/it/academy/task10/DAO/AdminDao.java
@@ -1,9 +1,9 @@
-package by.it.academy.task10.DAO2;
+package by.it.academy.task10.DAO;
 
 import by.it.academy.task10.entity.Admin;
 import by.it.academy.task10.util.HibernateUtil;
 
-public class AdminDao extends DAOBase<Admin>{
+public class AdminDao extends AbstractDAO<Admin> {
 
     public AdminDao() {
         super(Admin.class, HibernateUtil.getEntityManager());
