+++ repository/src/main/java/by/it/academy/task10/DAO/CourseDao.java
@@ -1,9 +1,9 @@
-package by.it.academy.task10.DAO2;
+package by.it.academy.task10.DAO;
 
 import by.it.academy.task10.entity.Course;
 import by.it.academy.task10.util.HibernateUtil;
 
-public class CourseDao extends DAOBase<Course> {
+public class CourseDao extends AbstractDAO<Course> {
 
     public CourseDao() {
         super(Course.class, HibernateUtil.getEntityManager());
