+++ repository/src/main/java/by/it/academy/task10/DAO/GenericDAO.java
@@ -1,11 +1,11 @@
-package by.it.academy.task10.dao;
+package by.it.academy.task10.DAO;
 
-import java.io.Serializable;
 import java.sql.SQLException;
 import java.util.List;
 
-public interface GenericDao<T> extends Serializable  {
-    T findOne(Object id) throws SQLException;
+public interface GenericDAO<T> {
+
+    T findOne(Integer id) throws SQLException;
 
     @SuppressWarnings("unchecked")
     List<T> findAll();
