+++ repository/src/main/java/by/it/academy/task10/DAO/MentorDao.java
@@ -1,9 +1,9 @@
-package by.it.academy.task10.DAO2;
+package by.it.academy.task10.DAO;
 
 import by.it.academy.task10.entity.Mentor;
 import by.it.academy.task10.util.HibernateUtil;
 
-public class MentorDao extends DAOBase<Mentor> {
+public class MentorDao extends AbstractDAO<Mentor> {
 
     public MentorDao() {
         super(Mentor.class, HibernateUtil.getEntityManager());
