+++ repository/src/main/java/by/it/academy/task10/DAO/StudentDao.java
@@ -1,9 +1,9 @@
-package by.it.academy.task10.DAO2;
+package by.it.academy.task10.DAO;
 
 import by.it.academy.task10.entity.Student;
 import by.it.academy.task10.util.HibernateUtil;
 
-public class StudentDao extends DAOBase<Student> {
+public class StudentDao extends AbstractDAO<Student> {
 
     public StudentDao() {
         super(Student.class, HibernateUtil.getEntityManager());
