+++ repository/src/main/java/by/it/academy/task10/DAO/TaskDao.java
@@ -1,9 +1,9 @@
-package by.it.academy.task10.DAO2;
+package by.it.academy.task10.DAO;
 
 import by.it.academy.task10.entity.Task;
 import by.it.academy.task10.util.HibernateUtil;
 
-public class TaskDao extends DAOBase<Task>{
+public class TaskDao extends AbstractDAO<Task> {
 
     public TaskDao() {
         super(Task.class, HibernateUtil.getEntityManager());
