+++ repository/src/main/resources/liquibase/changeLog/JD2Task10-changelog.xml
@@ -1,20 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
 <databaseChangeLog
         xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
-                      https://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.6.xsd">
-        
-    <changeSet id="jd2task10-1" author="JustShooter">
+                      http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.9.xsd"
+        objectQuotingStrategy="QUOTE_ONLY_RESERVED_WORDS">
+    <changeSet id="jd2Task10-1" author="JustShooter">
         <createTable tableName="course">
             <column autoIncrement="true" name="id" type="INT">
                 <constraints nullable="false" primaryKey="true" primaryKeyName="pk_course"/>
             </column>
             <column name="course_title" type="VARCHAR(255)"/>
+            <column name="mentor_id" type="INT"/>
         </createTable>
     </changeSet>
-    <changeSet id="jd2task10-2" author="JustShooter">
+    <changeSet id="jd2Task10-2" author="JustShooter">
         <createTable tableName="course_student">
             <column name="course_id" type="INT">
                 <constraints nullable="false" primaryKey="true" primaryKeyName="pk_course_student"/>
@@ -24,7 +24,7 @@
             </column>
         </createTable>
     </changeSet>
-    <changeSet id="jd2task10-3" author="JustShooter">
+    <changeSet id="jd2Task10-3" author="JustShooter">
         <createTable tableName="report">
             <column autoIncrement="true" name="id" type="INT">
                 <constraints nullable="false" primaryKey="true" primaryKeyName="pk_report"/>
@@ -36,7 +36,7 @@
             <column name="mentor_id" type="INT"/>
         </createTable>
     </changeSet>
-    <changeSet id="jd2task10-4" author="JustShooter">
+    <changeSet id="jd2Task10-4" author="JustShooter">
         <createTable tableName="task">
             <column autoIncrement="true" name="id" type="INT">
                 <constraints nullable="false" primaryKey="true" primaryKeyName="pk_task"/>
@@ -45,7 +45,7 @@
             <column name="course_id" type="INT"/>
         </createTable>
     </changeSet>
-    <changeSet id="jd2task10-5" author="JustShooter">
+    <changeSet id="jd2Task10-5" author="JustShooter">
         <createTable tableName="user">
             <column autoIncrement="true" name="id" type="INT">
                 <constraints nullable="false" primaryKey="true" primaryKeyName="pk_user"/>
@@ -54,36 +54,35 @@
             <column name="name" type="VARCHAR(255)"/>
             <column name="surname" type="VARCHAR(255)"/>
             <column name="role" type="VARCHAR(255)"/>
-            <column name="course_id" type="INT"/>
         </createTable>
     </changeSet>
-    <changeSet id="jd2task10-6" author="JustShooter">
+    <changeSet id="jd2Task10-6" author="JustShooter">
+        <addForeignKeyConstraint baseColumnNames="mentor_id" baseTableName="course" constraintName="FK_COURSE_ON_MENTOR"
+                                 referencedColumnNames="id" referencedTableName="user"/>
+    </changeSet>
+    <changeSet id="jd2Task10-7" author="JustShooter">
         <addForeignKeyConstraint baseColumnNames="mentor_id" baseTableName="report" constraintName="FK_REPORT_ON_MENTOR"
                                  referencedColumnNames="id" referencedTableName="user"/>
     </changeSet>
-    <changeSet id="jd2task10-7" author="JustShooter">
+    <changeSet id="jd2Task10-8" author="JustShooter">
         <addForeignKeyConstraint baseColumnNames="student_id" baseTableName="report"
                                  constraintName="FK_REPORT_ON_STUDENT" referencedColumnNames="id"
                                  referencedTableName="user"/>
     </changeSet>
-    <changeSet id="jd2task10-8" author="JustShooter">
+    <changeSet id="jd2Task10-9" author="JustShooter">
         <addForeignKeyConstraint baseColumnNames="task_id" baseTableName="report" constraintName="FK_REPORT_ON_TASK"
                                  referencedColumnNames="id" referencedTableName="task"/>
     </changeSet>
-    <changeSet id="jd2task10-9" author="JustShooter">
+    <changeSet id="jd2Task10-10" author="JustShooter">
         <addForeignKeyConstraint baseColumnNames="course_id" baseTableName="task" constraintName="FK_TASK_ON_COURSE"
                                  referencedColumnNames="id" referencedTableName="course"/>
     </changeSet>
-    <changeSet id="jd2task10-10" author="JustShooter">
-        <addForeignKeyConstraint baseColumnNames="course_id" baseTableName="user" constraintName="FK_USER_ON_COURSE"
-                                 referencedColumnNames="id" referencedTableName="course"/>
-    </changeSet>
-    <changeSet id="jd2task10-11" author="JustShooter">
+    <changeSet id="jd2Task10-11" author="JustShooter">
         <addForeignKeyConstraint baseColumnNames="course_id" baseTableName="course_student"
                                  constraintName="fk_coustu_on_course" referencedColumnNames="id"
                                  referencedTableName="course"/>
     </changeSet>
-    <changeSet id="jd2task10-12" author="JustShooter">
+    <changeSet id="jd2Task10-12" author="JustShooter">
         <addForeignKeyConstraint baseColumnNames="student_id" baseTableName="course_student"
                                  constraintName="fk_coustu_on_student" referencedColumnNames="id"
                                  referencedTableName="user"/>
